requirements.txt
cython>=0.29.0
Pillow>=3.0.0
numpy>=1.7.0
scipy>=1.0.0
scikit-image>=0.10.0